:green_heart: Clean some code.
