+++ src/logger.ts
@@ -1,0 +1,5 @@
+const log = (message: string) => {
+  console.log(new Date(), message);
+};
+
+export default log;
